+++ vdefs.h
@@ -25,8 +25,8 @@ typedef struct tagFreelist
 
 typedef struct tagPoint
     {
-    float x ;
-    float y ;
+    double x ;
+    double y ;
     } Point ;
 
 /* structure used both for sites and for vertices */
@@ -41,7 +41,7 @@ typedef struct tagSite
 
 typedef struct tagEdge
     {
-    float a, b, c ;
+    double a, b, c ;
     Site * ep[2] ;
     Site * reg[2] ;
     int edgenbr ;
@@ -58,7 +58,7 @@ typedef struct tagHalfedge
     int ELrefcnt ;
     char ELpm ;
     Site * vertex ;
-    float ystar ;
+    double ystar ;
     struct tagHalfedge * PQnext ;
     } Halfedge ;
 
@@ -84,16 +84,16 @@ Edge * bisect(Site *, Site *) ;
 Site * intersect(Halfedge *, Halfedge *) ;
 int right_of(Halfedge *, Point *) ;
 void endpoint(Edge *, int, Site *) ;
-float dist(Site *, Site *) ;
+double dist(Site *, Site *) ;
 void makevertex(Site *) ;
 void deref(Site *) ;
 void ref(Site *) ;
-extern float deltax, deltay ;
+extern double deltax, deltay ;
 extern int nsites, nedges, sqrt_nsites, nvertices ;
 extern Freelist sfl, efl ;
 
 /* heap.c */
-void PQinsert(Halfedge *, Site *, float) ;
+void PQinsert(Halfedge *, Site *, double) ;
 void PQdelete(Halfedge *) ;
 int PQbucket(Halfedge *) ;
 int PQempty(void) ;
@@ -105,11 +105,11 @@ extern Halfedge * PQhash ;
 
 /* main.c */
 extern int sorted, triangulate, plot, debug, nsites, siteidx ;
-extern float xmin, xmax, ymin, ymax ;
+extern double xmin, xmax, ymin, ymax ;
 extern Site * sites ;
 extern Freelist sfl ;
 extern AV *lines_out, *edges_out, *vertices_out;
-int compute_voronoi(Site *, int, float, float, float, float, int, AV*, AV*, AV*);
+int compute_voronoi(Site *, int, double, double, double, double, int, AV*, AV*, AV*);
 
 /* getopt.c */
 extern int getopt(int, char *const *, const char *);
@@ -123,9 +123,9 @@ void free_all(void);
 
 /* output.c */
 void openpl(void) ;
-void line(float, float, float, float) ;
-void circle(float, float, float) ;
-void range(float, float, float, float) ;
+void line(double, double, double, double) ;
+void circle(double, double, double) ;
+void range(double, double, double, double) ;
 void out_bisector(Edge *) ;
 void out_ep(Edge *) ;
 void out_vertex(Site *) ;
